+++ src/fzp.h
@@ -23,7 +23,7 @@ extern struct fzp *fzp_open(const char *path, const char *mode);
 extern struct fzp *fzp_gzopen(const char *path, const char *mode);
 extern int fzp_close(struct fzp **fzp);
 
-extern size_t fzp_read(struct fzp *fzp, void *ptr, size_t nmemb);
+extern int fzp_read(struct fzp *fzp, void *ptr, size_t nmemb);
 extern size_t fzp_write(struct fzp *fzp, const void *ptr, size_t nmemb);
 extern int fzp_eof(struct fzp *fzp);
 extern int fzp_flush(struct fzp *fzp);
